% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq100_approve.r
\name{approve_demo}
\alias{approve_demo}
\title{Check security to approve listed on nasdaq exchange demo}
\usage{
approve_demo()
}
\value{
List with approve to listed on exchange \code{security}
}
\description{
Check security to approve listed on nasdaq exchange demo
}
\examples{
approve_demo()
}
