% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approve.r
\docType{class}
\name{approve_message}
\alias{approve_message}
\title{Class providing object with methods for nasdaq security approve result}
\format{\code{\link{R6Class}} object.}
\usage{
approve_message
}
\value{
Object of \code{\link{R6Class}} with methods for nasdaq security approve result
}
\description{
Class providing object with methods for nasdaq security approve result
}
\section{Fields}{

}

\examples{
approve_message$new()
}
\keyword{data}
