% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq100.r
\docType{class}
\name{nasdaq100}
\alias{nasdaq100}
\title{Class providing object with methods for nasdaq security approve result}
\format{\code{\link{R6Class}} object.}
\usage{
nasdaq100
}
\value{
Object of \code{\link{R6Class}} with methods for nasdaq security approve result
}
\description{
Class providing object with methods for nasdaq security approve result
}
\section{Fields}{

}

\examples{
nasdaq100$new()
}
\keyword{data}
