% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq100_approve.r
\name{nasdaq_approve}
\alias{nasdaq_approve}
\title{Check security to approve listed on nasdaq exchange}
\usage{
nasdaq_approve(securities = NA)
}
\arguments{
\item{security}{A lsit.}
}
\value{
List with approve to listed on exchange \code{security}
}
\description{
Check security to approve listed on nasdaq exchange
}
\examples{
add(1, 1)
add(10, 1)
}
